Рефакторинг компонента LocationsTabs
