+++ project/src/components/locations-tabs/locations-tabs.tsx
@@ -1,12 +1,32 @@
+import { Dispatch } from 'react';
+import { connect, ConnectedProps } from 'react-redux';
+import { setCity } from '../../store/action';
+import { getSelectedCity } from '../../store/offers-list/selectors';
+import { Actions } from '../../types/store/actions';
+import { State } from '../../types/store/state';
 import { AvailableCity } from '../../utils/const';
 
-type LocationsTabsProps = {
-  selectedLocation: string,
-  locations: Array<AvailableCity>,
-  onSetLocation: (city: string) => void,
-}
+const availableCities = Object.values(AvailableCity);
 
-function LocationsTabs({selectedLocation, locations, onSetLocation}: LocationsTabsProps): JSX.Element {
+const mapStateToProps = (state: State) => ({
+  selectedCity: getSelectedCity(state),
+});
+
+const mapDispatchToProps = (dispatch: Dispatch<Actions>) => ({
+  onSetCity(city: string) {
+    dispatch(setCity(city));
+  },
+});
+
+const connector = connect(mapStateToProps, mapDispatchToProps);
+
+type PropsFromRedux = ConnectedProps<typeof connector>;
+type ConnectedComponentProps = PropsFromRedux;
+
+
+function LocationsTabs(props: ConnectedComponentProps): JSX.Element {
+  const {selectedCity, onSetCity} = props;
+
   const activeLocationClassList = 'locations__item-link tabs__item tabs__item--active';
   const locationClassList = 'locations__item-link tabs__item';
 
@@ -14,11 +34,11 @@ function LocationsTabs({selectedLocation, locations, onSetLocation}: LocationsTa
     <div className='tabs'>
       <section className='locations container'>
         <ul className='locations__list tabs__list'>
-          {locations.map((location) => (
+          {availableCities.map((location) => (
             <li className='locations__item' key={location}>
               <a
-                className={location.toLowerCase() === selectedLocation.toLowerCase() ? activeLocationClassList : locationClassList}
-                onClick={() => onSetLocation(location)}
+                className={location.toLowerCase() === selectedCity.toLowerCase() ? activeLocationClassList : locationClassList}
+                onClick={() => onSetCity(location)}
                 href='#temp'
               >
                 <span>{location}</span>
@@ -31,4 +51,5 @@ function LocationsTabs({selectedLocation, locations, onSetLocation}: LocationsTa
   );
 }
 
-export default LocationsTabs;
+export { LocationsTabs };
+export default connector(LocationsTabs);
